+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 新特性
 
 * [#101](https://github.com/TOMO-CAT/xmake/issues/101): 安装 local package 时展示日志
+* [#104](https://github.com/TOMO-CAT/xmake/issues/104): 删除了 project 的 indirect package 也会在 xmake build 的触发重新安装
 
 ## v3.0.3